Boss

+++ Strategies/Drift_Bina_ARB/Drift_Bina_ARB_Live.py
@@ -37,7 +37,7 @@ class Initialize:
 
     def __init__(self):
         self.LIMIT_DATA = True
-        self.ZSCORE_PERIOD = 6 * 3600 / 5  # Edit first number, hours of period (hours * minute in seconds / 5s data frequency)
+        self.ZSCORE_PERIOD = int(6 * 3600 / 5)  # Edit first number, hours of period (hours * minute in seconds / 5s data frequency)
         self.FAST_AVG = 24
         self.QUARTILE = 0.15
         self.MIN_REGULAR_GAP = 0.44
